backend/requirements.txt
Django==4.2.20
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
PyMySQL==1.1.1
channels==4.0.0
daphne==4.0.0
channels_redis==4.2.0
